// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  index: typeof index;
  insights: typeof insights;
  recordings: typeof recordings;
  streaming: typeof streaming;
  workflows: typeof workflows;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;